pyproject.toml
[project]
name = "dfhq"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.8.1",
    "bitsandbytes>=0.46.1",
    "diffusers>=0.34.0",
    "protobuf>=6.31.1",
    "python-dotenv>=1.1.1",
    "sentencepiece>=0.2.0",
    "transformers>=4.53.1",
    "xformers>=0.0.31.post1",
]

[dependency-groups]
dev = [
    "torchvision>=0.22.1",
]
